Move graph loader into common
